Improvements to Codegen technically finished
Codegen will now work without problems for 1.0.48
Additionally reworked the style of functions.hpp and
fixed various bugs:

 - only values added by supported extensions will be used now
 - extensions can add bitmask values
 - more space-efficient functions.hpp
 - temporarily disabled the version static_assert check
 - to be reintroduced later on (maybe sth like static_warn)

The changes to functions.hpp were made with the intentions that
the file is not actively used for understanding what happens
in the wrapper functions.
If this will turn out wrong the changes should be reversed.